pyproject.toml
[project]
name = "trepp-self-storage"
version = "0.1.0"
description = "Self-storage CMBS analysis: vacancy, revenue/NOI, and distress signals"
authors = [{name = "Giovanni Aiello"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "snowflake-connector-python>=3.0.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "catboost>=1.2.0",
    "lightgbm>=4.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-dotenv>=1.0.0",
    "shap>=0.43.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["src*"]
